requirements.txt
matplotlib
numpy
opencv-python
opencv_contrib_python
Pillow
pymeshlab
PyOpenGL
PyOpenGL_accelerate
PyYAML>=6.0
scipy
setuptools
scikit-image
termcolor
tqdm
joblib
yacs>=0.1.8
rembg
loguru
kornia
trimesh
pytorch_lightning
pytorch3d
filterpy
cython
pyrender
json_tricks
torchgeometry
opendr
flatten_dict
albumentations
numba
fastai
jpeg4py
git+https://github.com/scottandrews/chumpy.git
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
git+https://github.com/Project-Splinter/streamer_pytorch
git+https://github.com/Project-Splinter/human_det
git+https://github.com/giacaglia/pytube.git